+++ MPChartLib/src/main/java/com/github/mikephil/charting/data/PieEntry.java
@@ -12,38 +12,38 @@ public class PieEntry extends Entry {
 
     private String label;
 
-    public PieEntry(float value) {
+    public PieEntry(double value) {
         super(0f, value);
     }
 
-    public PieEntry(float value, Object data) {
-        super(0f, value, data);
+    public PieEntry(double value, Object data) {
+        super(0, value, data);
     }
 
-    public PieEntry(float value, Drawable icon) {
-        super(0f, value, icon);
+    public PieEntry(double value, Drawable icon) {
+        super(0, value, icon);
     }
 
-    public PieEntry(float value, Drawable icon, Object data) {
-        super(0f, value, icon, data);
+    public PieEntry(double value, Drawable icon, Object data) {
+        super(0, value, icon, data);
     }
 
-    public PieEntry(float value, String label) {
+    public PieEntry(double value, String label) {
         super(0f, value);
         this.label = label;
     }
 
-    public PieEntry(float value, String label, Object data) {
+    public PieEntry(double value, String label, Object data) {
         super(0f, value, data);
         this.label = label;
     }
 
-    public PieEntry(float value, String label, Drawable icon) {
+    public PieEntry(double value, String label, Drawable icon) {
         super(0f, value, icon);
         this.label = label;
     }
 
-    public PieEntry(float value, String label, Drawable icon, Object data) {
+    public PieEntry(double value, String label, Drawable icon, Object data) {
         super(0f, value, icon, data);
         this.label = label;
     }
@@ -53,7 +53,7 @@ public class PieEntry extends Entry {
      *
      * @return
      */
-    public float getValue() {
+    public double getValue() {
         return getY();
     }
 
@@ -67,14 +67,14 @@ public class PieEntry extends Entry {
 
     @Deprecated
     @Override
-    public void setX(float x) {
+    public void setX(double x) {
         super.setX(x);
         Log.i("DEPRECATED", "Pie entries do not have x values");
     }
 
     @Deprecated
     @Override
-    public float getX() {
+    public double getX() {
         Log.i("DEPRECATED", "Pie entries do not have x values");
         return super.getX();
     }
